Added a logo variation for dark mode
Previous logotype was black and did not appear when browsing in dark mode, the most common mode. I inverted the colors and added the result as an alternative image. Pure white seemed too bright so I lowered colors with FFMPEG to F2F2F2, close to what GitHub text is colored like.

I also took the liberty to replace alt text from "logo" to an empty string. As acknowledging the logo is of no benefit to screenreader users, it's better for the program to just ignore it. At least that's what I learned in an accessibility lecture.